# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Library Management API
  version: 1.0.0

servers:
  - url: http://localhost:5000

paths:
  /books:
    get:
      summary: List all books
      responses:
        '200':
          description: A list of books
    post:
      summary: Add a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - author
                - published_year
                - isbn
              properties:
                title:
                  type: string
                author:
                  type: string
                published_year:
                  type: integer
                isbn:
                  type: string
                genre:
                  type: string
      responses:
        '201':
          description: Book added successfully

  /books/search:
    get:
      summary: Search for books
      parameters:
        - name: author
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
        - name: genre
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Filtered list of books

  /books/{isbn}:
    delete:
      summary: Delete a book by ISBN
      parameters:
        - in: path
          name: isbn
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book deleted successfully
        '404':
          description: Book not found
    put:
      summary: Update a book by ISBN
      parameters:
        - in: path
          name: isbn
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                author:
                  type: string
                published_year:
                  type: integer
                genre:
                  type: string
      responses:
        '200':
          description: Book updated successfully
        '404':
          description: Book not found